+++ crates/cachet/src/fallback.rs
@@ -20,117 +20,10 @@ use crate::refresh::TimeToRefresh;
 use crate::telemetry::ext::ClockExt;
 use crate::telemetry::{CacheActivity, CacheOperation, CacheTelemetry};
 
-/// Type alias for promotion predicate functions.
-type PromotionPredicate<V> = Arc<dyn Fn(&CacheEntry<V>) -> bool + Send + Sync>;
-
-/// Policy for promoting values from fallback to primary cache.
-///
-/// When a cache miss occurs in the primary tier and a value is found in the
-/// fallback tier, the promotion policy determines whether to copy that value
-/// back to the primary tier for faster future access.
-///
-/// # Examples
-///
-/// ```
-/// use cachet::FallbackPromotionPolicy;
-///
-/// // Always promote (default)
-/// let policy = FallbackPromotionPolicy::<String>::always();
-///
-/// // Never promote
-/// let policy = FallbackPromotionPolicy::<String>::never();
-///
-/// // Promote based on a condition
-/// let policy = FallbackPromotionPolicy::<String>::when(|entry| entry.value().len() >= 5);
-/// ```
-#[derive(Debug, Default)]
-pub struct FallbackPromotionPolicy<V>(PolicyType<V>);
-
-#[derive(Default)]
-enum PolicyType<V> {
-    /// Always promote values to primary cache.
-    #[default]
-    Always,
-    /// Never promote values to primary cache.
-    Never,
-    /// Promote based on a boxed predicate that can capture state.
-    ///
-    /// Use this when you need to capture external state in the predicate.
-    /// Has slight overhead from dynamic dispatch.
-    When(PromotionPredicate<V>),
-}
-
-impl<V> std::fmt::Debug for PolicyType<V> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Always => write!(f, "Always"),
-            Self::Never => write!(f, "Never"),
-            Self::When(_) => write!(f, "WhenBoxed(<closure>)"),
-        }
-    }
-}
-
-impl<V> FallbackPromotionPolicy<V> {
-    /// Creates a policy that always promotes values to the primary cache.
-    ///
-    /// This is the default behavior and maximizes cache hit rates at the cost
-    /// of additional writes to the primary tier.
-    #[must_use]
-    pub fn always() -> Self {
-        Self(PolicyType::Always)
-    }
-
-    /// Creates a policy that never promotes values to the primary cache.
-    ///
-    /// Use this when the fallback tier is already fast enough and you want
-    /// to avoid write overhead to the primary tier.
-    #[must_use]
-    pub fn never() -> Self {
-        Self(PolicyType::Never)
-    }
-
-    /// Creates a policy using a predicate closure.
-    ///
-    /// The closure can capture external state if needed.
-    ///
-    /// ```no_run
-    /// use cachet::{Cache, CacheEntry, FallbackPromotionPolicy};
-    /// use tick::Clock;
-    ///
-    /// let min_len = 3;
-    /// let clock = Clock::new_tokio();
-    /// let l2 = Cache::builder::<String, String>(clock.clone()).memory();
-    /// let cache = Cache::builder::<String, String>(clock)
-    ///     .memory()
-    ///     .fallback(l2)
-    ///     .promotion_policy(FallbackPromotionPolicy::when(
-    ///         move |entry: &CacheEntry<String>| entry.value().len() >= min_len,
-    ///     ))
-    ///     .build();
-    /// ```
-    pub fn when<F>(predicate: F) -> Self
-    where
-        F: Fn(&CacheEntry<V>) -> bool + Send + Sync + 'static,
-    {
-        Self(PolicyType::When(Arc::new(predicate)))
-    }
-
-    /// Returns true if the response should be promoted to primary.
-    #[inline]
-    pub(crate) fn should_promote(&self, response: &CacheEntry<V>) -> bool {
-        match &self.0 {
-            PolicyType::Always => true,
-            PolicyType::Never => false,
-            PolicyType::When(pred) => pred(response),
-        }
-    }
-}
-
 pub(crate) struct FallbackCacheInner<K, V, P, F> {
     pub(crate) name: CacheName,
     pub(crate) primary: P,
     pub(crate) fallback: F,
-    pub(crate) policy: FallbackPromotionPolicy<V>,
     pub(crate) clock: Clock,
     pub(crate) refresh: Option<TimeToRefresh<K>>,
     pub(crate) telemetry: CacheTelemetry,
@@ -147,8 +40,8 @@ impl<K, V, P, F> std::fmt::Debug for FallbackCacheInner<K, V, P, F> {
 
 /// A two-tier cache that checks a primary tier, then falls back to a secondary tier.
 ///
-/// On a primary cache miss, the fallback tier is queried. Based on the promotion
-/// policy, successful fallback hits may be promoted back to the primary tier.
+/// On a primary cache miss, the fallback tier is queried. Successful fallback hits
+/// may be promoted back to the primary tier.
 ///
 /// Construct this via `Cache::builder().fallback()` rather than directly.
 ///
@@ -167,7 +60,6 @@ impl<K, V, P, F> std::fmt::Debug for FallbackCacheInner<K, V, P, F> {
 ///     .memory()
 ///     .ttl(Duration::from_secs(60))
 ///     .fallback(l2)
-///     .promotion_policy(FallbackPromotionPolicy::always())
 ///     .build();
 /// ```
 #[derive(Debug)]
@@ -181,7 +73,6 @@ impl<K, V, P, F> FallbackCache<K, V, P, F> {
         name: CacheName,
         primary: P,
         fallback: F,
-        policy: FallbackPromotionPolicy<V>,
         clock: Clock,
         refresh: Option<TimeToRefresh<K>>,
         telemetry: CacheTelemetry,
@@ -191,7 +82,6 @@ impl<K, V, P, F> FallbackCache<K, V, P, F> {
                 name,
                 primary,
                 fallback,
-                policy,
                 clock,
                 refresh,
                 telemetry,
@@ -220,9 +110,7 @@ where
         // Propagate any error from fallback
         let fallback_value = timed.result?;
 
-        if let Some(ref v) = fallback_value
-            && self.inner.policy.should_promote(v)
-        {
+        if let Some(ref v) = fallback_value {
             let timed_insert = self
                 .inner
                 .clock
@@ -323,12 +211,12 @@ mod tests {
         CacheWrapper::new("primary", MockCache::new(), clock, None, telemetry)
     }
 
-    fn make_fallback_cache(policy: FallbackPromotionPolicy<i32>) -> TestFallbackCache {
+    fn make_fallback_cache() -> TestFallbackCache {
         let clock = Clock::new_frozen();
         let primary = make_primary();
         let fallback_mock = MockCache::<String, i32>::new();
         let telemetry = TelemetryConfig::new().build();
-        FallbackCache::new("fallback", primary, fallback_mock, policy, clock, None, telemetry)
+        FallbackCache::new("fallback", primary, fallback_mock, clock, None, telemetry)
     }
 
     /// Tests that promotion from fallback to primary works correctly.
